requirements.txt
torch
transformers
flask
flask_cors
pillow
